admin posts

+++ database/migrations/2019_02_26_190244_create_user_roles_table.php
@@ -14,9 +14,9 @@ class CreateUserRolesTable extends Migration
     public function up()
     {
         Schema::create('user_roles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('role_id')->references('id')->on('roles')->onDelete('cascade');
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            $table->bigInteger('role_id')->references('id')->on('roles')->onDelete('cascade');
             $table->timestamps();
         });
     }
